[refactor] 클라이언트, 프리랜서 공통 컴포넌트 추출
